fix compile errorv5

+++ framework/src/passes/tensor_graph_pass/remove_redundant_reshape.cpp
@@ -25,11 +25,11 @@ using namespace npu::tile_fwk;
 namespace npu::tile_fwk {
 namespace {
 
-bool containsNegativeOne(const std::vector<int>& vec) {
+bool containsNegativeOne(const std::vector<int64_t>& vec) {
     if (vec.empty()) {
         return false;
     }
-    return std::find(vec.begin(), vec.end(), static_cast<int>(-1)) != vec.end();
+    return std::find(vec.begin(), vec.end(), static_cast<int64_t>(-1)) != vec.end();
 }
 
 Status CheckIOOperands(const Operation &op, LogicalTensorPtr &in, LogicalTensorPtr &out) {
